backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
alembic==1.13.3
python-multipart==0.0.9
Pillow==10.4.0
pytesseract==0.3.13
requests==2.32.3
scikit-learn==1.5.2
